chore(team): split linked run manager tests
